Various changes
The file "main_version.py" was modified. The main_version was updated from 1.3 to 1.4. The "import ctypes" and "import platform" lines were added. The temperature for the ChatCompletion.create function was changed from 0 to 0.5. The "get_keyboard_layout" function was added to solve an issue with keyboards not working properly. The "check_windows_version" function was added to get the Windows version. The "get_filepath" function was modified to use the correct keyboard layout for different regions. The "on_keypress" function was modified to use the correct keyboard shortcut for Windows 11.