Various changes
The main changes in this commit include updating the version to 1.6, importing the 'queue' and 'pystray' modules, and making changes to the code related to the system tray icon. The code now includes a custom icon and sets the current process ID. Additionally, there are some changes related to image handling. Some lines of code were removed, and some new lines were added. The cleanup code at the end of the file was also modified.